Haftalık commit, bir kaç servis, hata ayıklama ve bazı testler hariç backend tamamlandı, api dokümantasyonu cikarildi geriye front kaliyor

+++ src/main/java/com/anil/crm/repositories/CustomerRepository.java
@@ -1,7 +1,7 @@
 package com.anil.crm.repositories;
 
-import com.anil.crm.entities.Customer;
+import com.anil.crm.domain.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer,Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 }
