+++ src/main/java/com/anil/crm/repositories/TicketCommentRepository.java
@@ -1,7 +1,7 @@
 package com.anil.crm.repositories;
 
-import com.anil.crm.entities.TicketComment;
+import com.anil.crm.domain.TicketComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketCommentRepository extends JpaRepository<TicketComment,Integer> {
+public interface TicketCommentRepository extends JpaRepository<TicketComment, Long> {
 }
