+++ src/main/java/com/anil/crm/repositories/TicketRepository.java
@@ -1,7 +1,7 @@
 package com.anil.crm.repositories;
 
-import com.anil.crm.entities.Ticket;
+import com.anil.crm.domain.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepository extends JpaRepository<Ticket,Integer> {
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
 }
